Move Custom User Control views
