+++ ArchipelagoParser/Views/CustomUserControl/APIConnect.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.APIConnect"
+﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.CustomUserControl.APIConnect"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
