+++ ArchipelagoParser/Views/CustomUserControl/APIConnect.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace NoNiDev.ArchipelagoParser.Views.CustomUserControl
+{ 
+    /// <summary>
+    /// Interaction logic for APIConnect.xaml
+    /// </summary>
+    public partial class APIConnect : UserControl
+    {
+        public APIConnect()
+        {
+            InitializeComponent();
+        }
+    }
+}
