+++ ArchipelagoParser/Views/CustomUserControl/OpenFileButton.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.OpenFileButton"
+﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.CustomUserControl.OpenFileButton"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
