+++ ArchipelagoParser/Views/CustomUserControl/OpenFileButton.xaml.cs
@@ -1,0 +1,16 @@
+﻿using System.Windows.Controls;
+
+namespace NoNiDev.ArchipelagoParser.Views.CustomUserControl
+{
+    /// <summary>
+    /// Interaction logic for OpenFileButton.xaml
+    /// </summary>
+    public partial class OpenFileButton : UserControl
+    {
+        public OpenFileButton()
+        {
+            InitializeComponent();
+            DataContext = new ViewModel.OpenFileViewModel();
+        }
+    }
+}
