+++ ArchipelagoParser/Views/CustomUserControl/RandoStat/RandoSlotView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.RandoSlotView"
+﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.CustomUserControl.RandoStat.RandoSlotView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -14,8 +14,7 @@
         <conv:BooleanToHiddenVisibilityConverter x:Key="BooleanToHiddenVisibility" />
     </UserControl.Resources>
 
-    
-        <Grid HorizontalAlignment="Stretch">
+    <Grid HorizontalAlignment="Stretch">
         <Grid.RowDefinitions>
             <RowDefinition Height="*" />
         </Grid.RowDefinitions>
@@ -67,7 +66,6 @@
         <Image Grid.Column="6" 
         Source="{Binding Source={x:Static draw:SystemIcons.Warning},Converter={StaticResource IconToImageSource},Mode=OneWay}" ToolTip="Selectionner un joueur pour le slot"
        HorizontalAlignment="Center" Height="15" VerticalAlignment="Center" Width="15" Visibility="{Binding IsNameSelected, Converter={StaticResource BooleanToHiddenVisibility}}"/>
-  
 
     </Grid>
 </UserControl>
