+++ ArchipelagoParser/Views/CustomUserControl/RandoStat/RandoSlotView.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace NoNiDev.ArchipelagoParser.Views.CustomUserControl.RandoStat
+{
+    /// <summary>
+    /// Interaction logic for RandoSlotView.xaml
+    /// </summary>
+    public partial class RandoSlotView : UserControl
+    {
+        public RandoSlotView()
+        {
+            InitializeComponent();
+        }
+    }
+}
