+++ ArchipelagoParser/Views/CustomUserControl/SOHAR/SOHARPlayerLine.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.SOHARPlayerLine"
+﻿<UserControl x:Class="NoNiDev.ArchipelagoParser.Views.CustomUserControl.SOHAR.SOHARPlayerLine"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
