--- ArchipelagoParser/Views/OpenFileButton.xaml.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace NoNiDev.ArchipelagoParser.Views
-{
-    /// <summary>
-    /// Interaction logic for OpenFileButton.xaml
-    /// </summary>
-    public partial class OpenFileButton : UserControl
-    {
-        public OpenFileButton()
-        {
-            InitializeComponent();
-            DataContext = new ViewModel.OpenFileViewModel();
-        }
-    }
-}
