+++ ArchipelagoParser/Views/RandoStatsView.xaml
@@ -4,7 +4,9 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:draw="clr-namespace:System.Drawing;assembly=System.Drawing"
-                xmlns:conv="clr-namespace:NoNiDev.ArchipelagoParser.Converter"
+             xmlns:conv="clr-namespace:NoNiDev.ArchipelagoParser.Converter"
+             xmlns:uc="clr-namespace:NoNiDev.ArchipelagoParser.Views.CustomUserControl"
+             xmlns:ucRando="clr-namespace:NoNiDev.ArchipelagoParser.Views.CustomUserControl.RandoStat"
              xmlns:local="clr-namespace:NoNiDev.ArchipelagoParser.Views"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
@@ -25,7 +27,7 @@
             <!-- label requete ?<RowDefinition Height="auto" />-->
         </Grid.RowDefinitions>
         <!-- API -->
-        <local:APIConnect Grid.Row="0" 
+        <uc:APIConnect Grid.Row="0" 
                                       DataContext="{Binding RandoStatAPIVM}"/>
         <!-- HEADERS -->
         <Grid Grid.Row="1"
@@ -56,7 +58,7 @@
             <ListBox ItemsSource="{Binding Players}" HorizontalContentAlignment="Stretch">
                 <ListBox.ItemTemplate>
                     <DataTemplate>
-                        <local:RandoSlotView HorizontalContentAlignment="Stretch" />
+                        <ucRando:RandoSlotView HorizontalContentAlignment="Stretch" />
                     </DataTemplate>
                 </ListBox.ItemTemplate>
             </ListBox>
