+++ ArchipelagoParser/Views/SOHARView.xaml
@@ -5,6 +5,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:NoNiDev.ArchipelagoParser.Views"
              xmlns:uc="clr-namespace:NoNiDev.ArchipelagoParser.Views.CustomUserControl"
+             xmlns:ucSOHAR="clr-namespace:NoNiDev.ArchipelagoParser.Views.CustomUserControl.SOHAR"
              xmlns:vm="clr-namespace:NoNiDev.ArchipelagoParser.ViewModel"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
@@ -37,7 +38,7 @@
         <ListBox Grid.Column="0" ItemsSource="{Binding Players}" >
             <ListBox.ItemTemplate>
                 <DataTemplate>
-                    <local:SOHARPlayerLine HorizontalContentAlignment="Stretch" />
+                    <ucSOHAR:SOHARPlayerLine HorizontalContentAlignment="Stretch" />
                 </DataTemplate>
             </ListBox.ItemTemplate>
         </ListBox>
@@ -49,7 +50,7 @@
         <TextBlock Grid.Row="1" Grid.Column="0" Grid.ColumnSpan="2"
             Text="{Binding RequestResponse}"
             HorizontalAlignment="Center"/>-->
-        <local:APIConnect Grid.Row="0" Grid.Column="1"
+        <uc:APIConnect Grid.Row="0" Grid.Column="1"
                               DataContext="{Binding SoharAPIVM}"/>
         <Button Grid.Row="2" Grid.Column="0" 
                 Command="{Binding ButtonRC}" IsEnabled="{Binding IsEnabled}"
